+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     * Packages are always rebuilt, even if they haven't changed
     * Hardcoded planning of builds
     * Builds are sequential
+* Better environment variable support
+    * Each env var can now be provided from a file by appending it with `_FILE`
+      & passing the path to the file as value
 
 ## Fixed